pyproject.toml
[project]
name = "lp-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.1.1",
    "dash-iconify>=0.1.2",
    "dash-mantine-components>=2.1.0",
    "google-genai>=1.27.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pulp>=3.2.2",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.0",
    "sympy>=1.14.0",
]
